Download forecast with curl
The -G flag flag forces curl to make a GET request instead of a POST
request because curl won't do it correctly otherwise.  The server
returns a page saying forecast data is unavailable when
POST-requested.  Curl only sends the first parameter when doing a GET
request with the parameters appended to the url.  So we have to
specify parameters with -d and then force a GET instead of POST with -G.