Wrap iqpf calculation in a function
I'm not sure how far I should go with this function thing.  Like should
getting and parsing the XML be in a function?  Should figuring out how
many arguments there are and dealing with that be in a function?
(probably not because you'd have to pass it all the args)